Moving definitions and functions
This commit removes common definitions and utility functions from the
translator to a more convenient module.
